Move pod logic to separate library (#140)
